backend/requirements/base.txt
# Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.1

# Cache & Tasks
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Docs
drf-spectacular==0.27.0

# Encryption
cryptography==41.0.7

# Notifications
twilio==8.11.0

# File handling
Pillow==10.2.0
python-magic==0.4.27

# Rate limiting
django-ratelimit==4.1.0

# Filtering
django-filter==23.5

# Production server
gunicorn==21.2.0

# Logging
django-request-logging==0.7.5

# Utils
python-dateutil==2.8.2
